// src/models/operations/operations.enum.ts
export enum Operations {
  setupBatch = 'setupBatch',
  releaseBatch = 'releaseBatch',
  pushHatcheryData = 'pushHatcheryData',
  pushFarmerData = 'pushFarmerData',
  pushSlaughterhouseData = 'pushSlaughterhouseData',
  orderAnalysis = 'orderAnalysis',
  pushLaboratoryData = 'pushLaboratoryData',
  getBatch = 'getBatch',
  getProductionChain = 'getProductionChain',
  findNotarizations = 'findNotarizations',
}